Add a minimal fine-grained reactive library.
The library allows the user to define values (called "signals") whose
changes can be tracked by observer objects (called "effects"): whenever
a signal is updated, all its associated effects are executed.

The relationships between signals and effects are established
automatically: whenever an effect reads a signal it is automatically
bound to the signal and subsequently notified of any update of the
signal. So, there is no need to explicitly "subscribe" an effect as an
observer of a signal.

Besides, such bindings are dynamic and may change over time: each time
an effect is executed, its dependent signals are re-computed on the fly.
See the test file for an example.