Use a heuristic to detect potentially infinite recursion.
Detecting infinite recursion when cascading effects amounts to finding a
cycle in a DAG. As effects are propagated in a breadth-first manner,
that is not entirely straightforward (although it could be done), and
may slow down normal computation. Instead, use a heuristic and bail out
if the queue becomes too large.