Keep track of properties and allow cleaning them up.
Although a user might call Clear() on any defined property, when there
are many properties that becomes tedious. More importantly, there is no
way to clean up a cached property created with CreateMemo(), because
only the reader method of such properties is made accessible to the
clients. That may result in undesired triggering of lingering effects
that should have been removed from the cached properties.