Implement libreactive's pools.
We need a way for a client to clear the effects of all the properties
that the client has defined, including the hidden ones created with
CreateMemo(). To do that, we need to keep track of every property that
a given client has created. Pools are the way such bookkeeping is
implemented.